+++ examples/mplus/src/lib.rs
@@ -1,2 +1,0 @@
-pub static MPLUS_1P_REGULAR: &[u8] = include_bytes!("../fonts/mplus-1p-regular.ttf");
-pub static MPLUS_1M_REGULAR: &[u8] = include_bytes!("../fonts/mplus-1m-regular.ttf");